requirements.txt
mido==1.3.0
MouseInfo==0.1.3
packaging==23.2
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-rtmidi==1.5.8
pytweening==1.0.7
PyYAML==6.0.1
rubicon-objc==0.4.7
AppKit
